🎉 New Template
- mirko-orm: TypeScript ORM for Node.js based on Data Mapper
