+++ lib/mikro-orm/package.json
@@ -2,10 +2,10 @@
   "name": "typescript-express-starter",
   "version": "0.0.0",
   "description": "TypeScript + MikroORM + MongoDB + Express API Server",
-  "author": "tamzidkarim",
+  "author": "",
   "license": "ISC",
   "scripts": {
-    "start": "npm run build:tsc && cross-env NODE_ENV=production node dist/server.js",
+    "start": "npm run build && cross-env NODE_ENV=production node dist/server.js",
     "dev": "cross-env NODE_ENV=development nodemon",
     "build": "swc src -d dist --source-maps --copy-files",
     "build:tsc": "tsc && tsc-alias",
@@ -65,6 +65,7 @@
     "node-config": "^0.0.2",
     "node-gyp": "^8.1.0",
     "nodemon": "^2.0.9",
+    "pm2": "^5.1.0",
     "prettier": "^2.3.2",
     "supertest": "^6.2.2",
     "ts-jest": "^27.1.4",
