+++ lib/mikro-orm/src/controllers/auth.controller.ts
@@ -1,8 +1,8 @@
 import { NextFunction, Request, Response } from 'express';
 import { CreateUserDto } from '@dtos/users.dto';
 import { RequestWithUser } from '@interfaces/auth.interface';
+import { User } from '@interfaces/users.interface';
 import AuthService from '@services/auth.service';
-import { User } from '@/entities';
 
 class AuthController {
   public authService = new AuthService();
