+++ lib/mikro-orm/src/controllers/users.controller.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from 'express';
 import { CreateUserDto } from '@dtos/users.dto';
+import { User } from '@interfaces/users.interface';
 import userService from '@services/users.service';
-import { User } from '@/entities';
 
 class UsersController {
   public userService = new userService();
