--- lib/mikro-orm/src/entities/index.ts
@@ -1,2 +1,0 @@
-export { BaseEntity } from './BaseEntity';
-export { User } from './User';
