+++ lib/mikro-orm/src/interfaces/auth.interface.ts
@@ -1,5 +1,5 @@
-import { User } from '@/entities';
 import { Request } from 'express';
+import { User } from '@interfaces/users.interface';
 
 export interface DataStoredInToken {
   _id: string;
