+++ lib/mikro-orm/src/interfaces/users.interface.ts
@@ -1,0 +1,5 @@
+export interface User {
+  id: string;
+  email: string;
+  password: string;
+}
