+++ lib/mikro-orm/src/services/auth.service.ts
@@ -1,12 +1,12 @@
 import { hash, compare } from 'bcrypt';
 import { sign } from 'jsonwebtoken';
 import { SECRET_KEY } from '@config';
+import { DI } from '@databases';
 import { CreateUserDto } from '@dtos/users.dto';
 import { HttpException } from '@exceptions/HttpException';
 import { DataStoredInToken, TokenData } from '@interfaces/auth.interface';
-import { DI } from '@/databases';
+import { User } from '@interfaces/users.interface';
 import { isEmpty } from '@utils/util';
-import { User } from '@/entities';
 
 class AuthService {
   public async signup(userData: CreateUserDto): Promise<User> {
